refactor: Step 1 코드리뷰 반영
- 카카오 로그인 POST -> GET
- @JsonNaming 반영으로 인한 개별 필드 어노테이션 삭제
