+++ README.md
@@ -19,7 +19,7 @@
 
 ### Step 2
 
-* [ ] Step 1 코드리뷰 반영
+* [X] Step 1 코드리뷰 반영
 * [ ] 스키마 & 도메인 구현
 * [ ] 레포지토리 구현
 * [ ] 클라이언트 구현
