--- src/main/java/gift/dto/member/KakaoLoginRequest.java
@@ -1,5 +1,0 @@
-package gift.dto.member;
-
-public record KakaoLoginRequest(String code) {
-
-}
