+++ src/main/java/gift/external/kakao/token/dto/KakaoTokenResponse.java
@@ -1,18 +1,14 @@
-package gift.external.kakao.dto;
+package gift.external.kakao.token.dto;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public record KakaoTokenResponse(
-        @JsonProperty("access_token")
         String accessToken,
 
-        @JsonProperty("token_type")
         String tokenType,
 
-        @JsonProperty("expires_in")
         int expiresIn
 ) {
