+++ src/test/java/gift/controller/user/AuthControllerTest.java
@@ -1,5 +1,6 @@
 package gift.controller.user;
 
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -99,26 +100,21 @@ class AuthControllerTest {
     }
 
     @Test
-    @DisplayName("POST /api/members/login/kakao - 카카오 로그인 성공 (200)")
+    @DisplayName("GET /api/members/login/kakao - 카카오 로그인 성공 (200)")
     void kakaoLoginSuccess() throws Exception {
         String code = "auth-code";
         AuthResponse resp = new AuthResponse("kakao-jwt-token");
 
         Mockito.when(memberService.kakaoLogin(code)).thenReturn(resp);
 
-        mockMvc.perform(post("/api/members/login/kakao")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("""
-                                    {
-                                        "code": "auth-code"
-                                    }
-                                """))
+        mockMvc.perform(get("/api/members/login/kakao")
+                        .param("code", code))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.token").value("kakao-jwt-token"));
     }
 
     @Test
-    @DisplayName("POST /api/members/login/kakao - 실패 시 401")
+    @DisplayName("GET /api/members/login/kakao - 실패 시 401")
     void kakaoLoginFailure() throws Exception {
         String code = "bad-code";
 
@@ -126,13 +122,8 @@ class AuthControllerTest {
                 .thenThrow(
                         new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Invalid Kakao Code"));
 
-        mockMvc.perform(post("/api/members/login/kakao")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("""
-                                    {
-                                        "code": "bad-code"
-                                    }
-                                """))
+        mockMvc.perform(get("/api/members/login/kakao")
+                        .param("code", code))
                 .andExpect(status().isUnauthorized());
     }
 }
